% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Delta.R
\name{Delta_modi}
\alias{Delta_modi}
\title{Delta_modi}
\usage{
Delta_modi(x, y, loc_age, lambda)
}
\arguments{
\item{x}{Covariate}

\item{y}{response}

\item{loc_age}{location of the age in the covariate matrix}

\item{lambda}{regularizer for the stability of the algorithm}
}
\value{
geometric median of planar shape
}
\description{
Delta_modi
}
