% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMED.R
\name{LMED}
\alias{LMED}
\title{LMED}
\usage{
LMED(x, y, age_loc, lambda)
}
\arguments{
\item{x}{Covariate}

\item{y}{response}

\item{lambda}{regularizer for the stability of the algorithm}

\item{loc_age}{location of the age in the covariate matrix}
}
\value{
geometric median of planar shape
}
\description{
LMED
}
